requirements.txt
django
django-import-export